% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.r
\name{algo}
\alias{algo}
\title{Main algorithm}
\usage{
algo(
  quantile_number,
  input_dir = getwd(),
  output_dir = NULL,
  data = "data.csv",
  diagnostically_related = NULL,
  binary = NULL,
  continuous = NULL,
  n = NULL,
  causal_relationships = NULL,
  combinations = NULL,
  p_threshold = 0.05
)
}
\arguments{
\item{quantile_number}{The number of quantiles or strata to use.}

\item{input_dir}{The directory where the data is stored, defaults to the current working directory.}

\item{output_dir}{The directory where the output should be stored, defaults to NULL meaning no output is saved.}

\item{data}{The name of the data file or an input data.table, defaults to "data.csv". Column names need to be in a specific format; see README page for details.}

\item{diagnostically_related}{The file name of the diagnostically related variables, or a data frame with two columns. No column headers required; see README page for details. Defaults to NULL meaning no diagnostically related variables are provided. Parameter used for preprocessing step.}

\item{binary}{The file name of the binary combinations, or a data frame with three columns (EXP, STRAT, CASE_STATUS). Column headers required. Defaults to NULL meaning the combinations will be automatically generated. Parameter used for preprocessing step.}

\item{continuous}{The file name of the continuous combinations, or a data frame with three columns (EXP, STRAT, OUTCOME). Column headers required. Defaults to NULL meaning the combinations will be automatically generated. Parameter used for preprocessing step.}

\item{n}{The number of combinations in the dataset (used for calculating the Bonferroni threshold), defaults to NULL meaning the number of observations will be automatically calculated.}

\item{causal_relationships}{The file name of the causal relationships, or a data frame with two columns defining a causal relationship where C1 is known to cause C2 (C1 -> C2). Headers are required (named C1 and C2). This defines the causal and collider relationships between C and X in the actual algorithm. If NULL, or there is no entry for C1 -> C2, then the causal relationship is unknown. Defaults to NULL meaning no causal relationships are provided.}

\item{combinations}{Skip preprocessing and pass in data frames (combinations$binary and combinations$continuous) returned from the preprocess function. Defaults to NULL meaning it will preprocess the provided data.}

\item{p_threshold}{The threshold for statistical significance, defaults to 0.05.}
}
\value{
A list containing two data tables, one for binary results and one for continuous results. See README page for details.
}
\description{
Main algorithm
}
