% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.r
\name{generate_sample_data}
\alias{generate_sample_data}
\title{Generate sample data}
\usage{
generate_sample_data(output_dir)
}
\arguments{
\item{output_dir}{The directory where the sample data files will be stored}
}
\value{
Writes the following files to the specified output directory:
\itemize{
\item data.csv: A data frame with 4 exposures, 4 PRS, 4 stratifying variables, 1 continuous outcome,
and 1 binary outcome for 100K individuals. This setup results in 32 initial combinations.
\item diagnostically_related.txt: A tab-delimited text file containing diagnostically-related variables
for any exposure or stratifying variable associated with the outcome.
\item continuous_comb_interest.txt: A tab-delimited text file with columns EXP, STRAT, and OUTCOME containing
combinations of interest with continuous outcomes.
\item binary_comb_interest.txt: A tab-delimited text file with columns EXP, STRAT, and CASE_STATUS containing
combinations of interest with binary outcomes.
\item causal_relationships.txt: A tab-delimited text file with columns C1 and C2, where C1 is known to cause
C2.
}
}
\description{
Generate sample data
}
