% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.r
\name{mr}
\alias{mr}
\title{Stratified Mendelian Randomization}
\usage{
mr(
  method,
  exposure_name,
  outcome_name,
  stratifying_name,
  quantile_number,
  data,
  input_dir = getwd(),
  output_dir = NULL
)
}
\arguments{
\item{method}{Either "by_x", "by_g", "resid_strat", "raw_strat", or "resid_wald"}

\item{quantile_number}{The number of quantiles or strata to use}

\item{data}{The name of the file containing the data or a data.table object. Column names need to be in a specific format; see README page for details.}

\item{input_dir}{The directory where the data is stored}

\item{output_dir}{The directory where the output should be stored, defaults to NULL meaning the results will not be saved}

\item{x}{The exposure name}

\item{y}{The outcome name}

\item{c}{The stratifying variable name}
}
\description{
Stratified Mendelian Randomization
}
