% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.r
\name{preprocess}
\alias{preprocess}
\title{Preprocess data for stratified MR}
\usage{
preprocess(
  input_dir = getwd(),
  output_dir = NULL,
  data = "data.csv",
  diagnostically_related = NULL,
  binary = NULL,
  continuous = NULL
)
}
\arguments{
\item{input_dir}{The directory where the input data is stored, defaults to the current working directory.}

\item{output_dir}{The directory where the output data will be stored, defaults to NULL meaning that the output will not be saved.}

\item{data}{The name of the file containing the data, or a data.table object, defaults to "data.csv". Column names need to be in a specific format; see README page for details.}

\item{diagnostically_related}{The name of the file containing the diagnostically related pairs or a data.table object, defaults to NULL meaning that no diagnostically related pairs are provided. No header required; see README page for details.}

\item{binary}{The file name of the binary combinations, or a data frame with three columns (EXP, STRAT, CASE_STATUS). Column headers required. Defaults to NULL meaning the combinations will be automatically generated.}

\item{continuous}{The file name of the continuous combinations, or a data frame with three columns (EXP, STRAT, OUTCOME). Column headers required. Defaults to NULL meaning the combinations will be automatically generated.}
}
\description{
Preprocess data for stratified MR
}
